chore: add add method. Move data store tests
